// src/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: any; output: any };
};

export type AffectedRowsOutput = {
  __typename?: "AffectedRowsOutput";
  count: Scalars["Int"]["output"];
};

export type AggregateTenant = {
  __typename?: "AggregateTenant";
  _count?: Maybe<TenantCountAggregate>;
  _max?: Maybe<TenantMaxAggregate>;
  _min?: Maybe<TenantMinAggregate>;
};

export type AggregateUser = {
  __typename?: "AggregateUser";
  _count?: Maybe<UserCountAggregate>;
  _max?: Maybe<UserMaxAggregate>;
  _min?: Maybe<UserMinAggregate>;
};
// ...
export type Query = {
  __typename?: "Query";
  aggregateTenant: AggregateTenant;
  aggregateUser: AggregateUser;
  findFirstTenant?: Maybe<Tenant>;
  findFirstTenantOrThrow?: Maybe<Tenant>;
  findFirstUser?: Maybe<User>;
  findFirstUserOrThrow?: Maybe<User>;
  getTenant?: Maybe<Tenant>;
  getUser?: Maybe<User>;
  groupByTenant: Array<TenantGroupBy>;
  groupByUser: Array<UserGroupBy>;
  tenant?: Maybe<Tenant>;
  tenants: Array<Tenant>;
  user?: Maybe<User>;
  users: Array<User>;
};
// ...
export type Tenant = {
  __typename?: "Tenant";
  _count?: Maybe<TenantCount>;
  admin: User;
  adminId: Scalars["String"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  users: Array<User>;
};

export type TenantUsersArgs = {
  cursor?: InputMaybe<UserWhereUniqueInput>;
  distinct?: InputMaybe<Array<UserScalarFieldEnum>>;
  orderBy?: InputMaybe<Array<UserOrderByWithRelationInput>>;
  skip?: InputMaybe<Scalars["Int"]["input"]>;
  take?: InputMaybe<Scalars["Int"]["input"]>;
  where?: InputMaybe<UserWhereInput>;
};

export type TenantCount = {
  __typename?: "TenantCount";
  users: Scalars["Int"]["output"];
};
// ...
export type TenantGroupBy = {
  __typename?: "TenantGroupBy";
  _count?: Maybe<TenantCountAggregate>;
  _max?: Maybe<TenantMaxAggregate>;
  _min?: Maybe<TenantMinAggregate>;
  adminId: Scalars["String"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  id: Scalars["String"]["output"];
  name: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type User = {
  __typename?: "User";
  _count?: Maybe<UserCount>;
  adminOfTenants: Array<Tenant>;
  createdAt: Scalars["DateTime"]["output"];
  email: Scalars["String"]["output"];
  id: Scalars["ID"]["output"];
  name?: Maybe<Scalars["String"]["output"]>;
  providers: Array<UserProvider>;
  superAdmin: Scalars["Boolean"]["output"];
  tenants: Array<Tenant>;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type UserGroupBy = {
  __typename?: "UserGroupBy";
  _count?: Maybe<UserCountAggregate>;
  _max?: Maybe<UserMaxAggregate>;
  _min?: Maybe<UserMinAggregate>;
  createdAt: Scalars["DateTime"]["output"];
  email: Scalars["String"]["output"];
  id: Scalars["String"]["output"];
  name?: Maybe<Scalars["String"]["output"]>;
  password?: Maybe<Scalars["String"]["output"]>;
  superAdmin: Scalars["Boolean"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type GetTenantQueryVariables = Exact<{
  tenantId?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type GetTenantQuery = {
  __typename?: "Query";
  tenant?: { __typename?: "Tenant"; id: string; name: string; adminId: string } | null;
};
// ...
export const GetTenantDocument = gql`
  query getTenant($tenantId: String = "") {
    tenant(where: { id: $tenantId }) {
      id
      name
      adminId
    }
  }
`;